Fixed grid representation | Added selection of 2D grid of the 3D grid to reduce noise.
Fixed some logic errors in the creation of the picross grid causing the blocks to be created in the wrong places making physical representation and array representation mismatch.

Added 2D grid selection of the 3D grid in X, Y and Z axis as well as the ability to cycle between them. Method in-game to do this not yet implemented, but the functions are there.