Completely remade the TArray3D, will also rename the header and source files soon.
The problem with using a wrapper class for TArray is that you lose reflection completely, leaving you with the problem of keeping the garbage collector from removing uobjects in the array.

The solution for now is to have the user create the TArray themselves, letting them add UPROPERTY to it to make the garbage collector aware of it. The user then uses the functions in the namesapce Array3D to convert from 3D to 1D and 1D to 3D Indexes. All of the functions are non-member functions and thus the user needs to supply the dimensions to every function. The user therefore needs a separate variable storing the dimensions, which is the biggest loss here, since you no longer have a 3D array, you have an array that you yourself calculate 3D indexes for.

Array safety is also lower since nothing stops the user from resizing and otherwise manipulating the array in a way that doesn't conform to the 3D limitations.