Went back to saving the Instance Index for the Block.
Pro: We don't need to do a find operation every single time we need to get the InstanceIndex, when doing operations that affect a lot of blocks this can theoretically save us a lot of time.

Con: We need to keep an eagle eye on where we modify HISMs, whenever we remove an instance it may or may not swap this instace with the instance at the end to make removal faster, the block it swaps with will then have an outdated InstanceIndex which we need to update. (Line 332 - 339 in PicrossGrid.cpp)