Redid logic for ideal pawn transform | Added the ability to toggle input modes.
Previously we calculated the origin of the grid with the Unreal function, this results in a logical race condition however since the remove/add of HISM instances seems to be delayed past the point where we try to calculate the origin causing the origin to sometimes be incorrect. Solved this with a constant value for the local height of origin (corrected for asymetric grids) since the height of the grid is always constrained to the same hardcoded height.
Pros: Works
Cons: Hardcoding.